BJS2-5392: add changes after review

+++ src/main/java/faang/school/godbless/sprint2/BJS2_5392/Calculator.java
@@ -2,5 +2,5 @@ package faang.school.godbless.sprint2.BJS2_5392;
 
 @FunctionalInterface
 public interface Calculator {
-    int calculate(int a, int b);
+    long calculate(long a, long b);
 }
